+++ files/gitlab-cookbooks/gitlab/attributes/default.rb
@@ -641,6 +641,8 @@ default['gitlab']['nginx']['server_names_hash_bucket_size'] = 64
 # HSTS
 default['gitlab']['nginx']['hsts_max_age'] = 31536000
 default['gitlab']['nginx']['hsts_include_subdomains'] = false
+# Compression
+default['gitlab']['nginx']['gzip_enabled'] = true
 
 ###
 # Nginx status