+++ config/software/gitlab-monitor.rb
@@ -14,7 +14,7 @@
 # limitations under the License.
 #
 
-version = Gitlab::Version.new('gitlab-monitor', '2.5.0')
+version = Gitlab::Version.new('gitlab-monitor', '2.13.0')
 
 default_version version.print