+++ CHANGELOG.md
@@ -20,6 +20,7 @@ omnibus-gitlab repository.
 - Add option to disable healthcheck for storagedriver in registry
 - Restart runsv when log directory is changed 0a784647b
 - Bump rsync to 3.1.3 f539aa946
+- Patch bzip2 against CVE-2016-3189 552730bfa
 
 10.7.0